+++ src/config/menu.ts
@@ -78,43 +78,43 @@ const menus = [
   {
     title: 'Collections',
     path: '/collection',
-    type: 'staffManagement',
+    type: 'management',
     roles: ['Staff', 'Manager', 'Admin', 'Dynamic'],
   },
   {
     title: 'Role Management',
     path: '/role',
-    type: 'staffManagement',
+    type: 'management',
     roles: ['Manager', 'Admin'],
   },
   {
     title: 'Staff Management',
     path: '/staff-management',
-    type: 'staffManagement',
+    type: 'management',
     roles: ['Manager', 'Admin'],
   },
   {
     title: 'Contact management',
     path: '/contact-management',
-    type: 'staffManagement',
-    roles: ['Glific_admin'],
+    type: 'management',
+    roles: ['Glific_admin', 'Admin'],
   },
   {
     title: 'Organizations',
     path: '/organizations',
-    type: 'staffManagement',
+    type: 'management',
     roles: ['Glific_admin'],
   },
   {
     title: 'Consulting',
     path: '/consulting-hours',
-    type: 'staffManagement',
+    type: 'management',
     roles: ['Glific_admin'],
   },
   {
     title: 'Blocked Contacts',
     path: '/blocked-contacts',
-    type: 'staffManagement',
+    type: 'management',
     roles: ['Staff', 'Manager', 'Admin', 'Dynamic'],
   },
   {
