+++ src/containers/Form/FormLayout.tsx
@@ -22,6 +22,15 @@ import { organizationHasDynamicRole } from 'common/utils';
 import { getUserRole } from 'context/role';
 import styles from './FormLayout.module.css';
 
+export const Heading = ({ icon, formTitle }: any) => (
+  <Typography variant="h5" className={styles.Title}>
+    <IconButton disabled className={styles.Icon}>
+      {icon}
+    </IconButton>
+    {formTitle}
+  </Typography>
+);
+
 export interface FormLayoutProps {
   match: any;
   deleteItemQuery: DocumentNode;
@@ -615,15 +624,7 @@ export const FormLayout: React.SFC<FormLayoutProps> = ({
     formTitle = `Add a new ${listItemName}`; // case when adding a new item
   }
 
-  let heading = (
-    <Typography variant="h5" className={styles.Title}>
-      <IconButton disabled className={styles.Icon}>
-        {icon}
-      </IconButton>
-      {formTitle}
-    </Typography>
-  );
-
+  let heading = <Heading icon={icon} formTitle={formTitle} />;
   if (advanceSearch) {
     const data = advanceSearch({});
     if (data && data.heading) heading = data.heading;
